Marked di used and fixed xml warnings.

+++ DynamicPrice.Core/Services/FindProductsToReduceService.cs
@@ -33,6 +33,7 @@ public class FindProductsToReduceService : BackgroundService
 		});
 	private readonly ConcurrentDictionary<int, DateTime> _lastMonitorEnd = new();
 
+	// todo: di
 	public FindProductsToReduceService(
 		IServiceProvider serviceProvider,
 		IConfiguration config,
@@ -48,6 +49,7 @@ public class FindProductsToReduceService : BackgroundService
 		{
 			while (!token.IsCancellationRequested)
 			{
+				// todo: di
 				using var scope = _serviceProvider.CreateScope();
 				var context = scope.ServiceProvider.GetRequiredService<DynamicPriceCoreContext>();
 				var publishEndpoint = scope.ServiceProvider.GetRequiredService<IPublishEndpoint>();
@@ -88,10 +90,11 @@ public class FindProductsToReduceService : BackgroundService
 		}
 	}
 
+	// todo: di
 	private async IAsyncEnumerable<Product> FindProductsToReduceAsync(
-	DynamicPriceCoreContext context,
-	CancellationToken token,
-	int? productsCount = null)
+		DynamicPriceCoreContext context,
+		CancellationToken token,
+		int? productsCount = null)
 	{
 		var activeCompaniesIds = await context.ActiveCompanies
 			.Select(ac => ac.CompanyId)
@@ -102,7 +105,7 @@ public class FindProductsToReduceService : BackgroundService
 
 		var priceRulesActiveCompaniesQuery = context.PriceRules
 			.AsNoTracking()
-			.Where(pr => activeCompaniesIds.Contains((int)pr.CompanyId));
+			.Where(pr => activeCompaniesIds.Contains(pr.CompanyId));
 
 
 		var productsActiveCompaniesQuery = context.Products
